interp - Interpolate

Interpolate Cumulative Loss Development Factors

Usage

interp(new_age, cdf_array, age_array, cutoff = 450, method = 3)

interp.dblexp(new_age, age_high, age_low, cdf_high, cdf_low)

interp.exp(new_age, age_high, age_low, cdf_high, cdf_low)

interp.linear(new_age, age_high, age_low, cdf_high, cdf_low)

Arguments

new_age

integer - value of the new age whose CDF is to be interpolated

cdf_array

numeric vector of CDFs (usually representative of the selected factors)

age_array

numeric vector of ages corresponding to the supplied cdf_array

cutoff

the largest possible age, after which, no interpolation is performed

method

integer - must be 1, 2, or 3 where 1 represents linear, 2 represents exponential, and 3 represents double exponential. Defaults to 3, but falls back onto 1 if necessary.

age_low, age_high

Low and High ages

cdf_low, cdf_high

Low and High CDFs

Value

derived numeric value for the supplied new_age's CDF

Details

Actuaries often have to interpolate values in between the selected LDFs/CDFs in order to derive development factors at a variety of possible ages of maturity.

This generic function comes with three possible methods:

Linear Interpolation

Exponential Interpolation

Double Exponential Interpolation

Functions

interp.dblexp(): Double Exponential Interpolation

interp.exp(): Exponential Interpolation

interp.linear(): Linear Interpolation

Examples

cdfs <- c(3.579, 2.866, 2.489, 2.121, 1.876, 1.543, 1.222, 1.150, 1.109, 1.005, 1.0025)
ages <- seq(from = 12, to = (length(cdfs) * 12), by = 12)

interp(14, cdfs, ages)
#> [1] 3.438518
interp(12, cdfs, ages) == cdfs[[1]]
#> [1] TRUE
interp(27, cdfs, ages, method = 2)
#> [1] 2.758217